Drop Requires: line from fishsound.pc.in
This line is only needed if packages that build against libfishsound
needs access to the Cflags of the required packages. In this case, the
listed packages (vorbis, speex, flac) are precisely the packages that
libfishsound provide a wrapper API for, so apps compiling against
libfishsound really should not need their headers directly.

This was not a problem in 0.9.1, but in 1.0.0 the FLAC headers are
exported, and this include a really regrettably misnamed assert.h, which
shadows the system assert.h and causes compilation of sonic-visualiser
to fail.

I've tried removing the Requires: line and can report that it results in
a functioning sonic-visualiser build.

Reported-by: hircus